Update patch set 2
Patch Set 2:

MIPS port of Android existed from cupcake time frame and had been carrying kernel headers of 2.6.27 since then. The other kernel headers in the repository like linux/ are from the same time frame.

I will be able to update the patch with a later version of the kernel headers if that is what is desired.

Patch-set: 2