Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I'm not sure I think this is a good idea, because it's going to be very difficult to track this properly in the future, would we need to update the header later. At a minimum, I would put a huge comment that explains everything, and use a #ifdef __KERNEL__ ... old definitions .. #else ... new one #endif,
and commit this to the upstream kernel headers.

After all, that's exactly what the __KERNEL__ feature macro is supposed to be used for.

Alternatively, there are already a few types that the C library treats specially, and this is typically done with macro magic like, in <sys/foo.h>, we would do something like:

  /* forces the declaration of 'foo' to be for the '__kernel_foo' type instead */
  #define foo  __kernel_foo
  #include <asm/foo.h>
  #undef foo

  ...
  struct foo {
      // user-land definition
  };

I think either method would be preferable to a direct change like this.

Patch-set: 1
Label: Code-Review=-1