Update patch set 1
Patch Set 1:

Thanks for enumerating different ways of handling an issue
like this. I will use it as a reference for any future issues.

After some discussion internally in our team, we concluded that this patch is really not required.
 
The registers used for ptrace system call, are very arch specific and on mips they are different from pt_regs structure. So for fixing debuggerd, we are using a local definition of the structure, just the way x86 is currently doing.

Abandoning this patch for the same reason.

Patch-set: 1