Update patch set 2
Patch Set 2:

that commit seems to be when some of the can files were modified. But you also include socket.h. I would prefer if you would just take the files from the commit v3.3 so that it's clear which version of socket.h, etc we get.

Patch-set: 2
Label: Code-Review=0
Label: Verified=0