Update patch set 2
Patch Set 2:

We do not want to keep hand modified versions of any headers here. The headers here should be direct copies from upstream. This is why I suggested to just take the latest copies directly from v3.3 and be done with it. If some of them haven't changed  since the commit you mentioned, that's fine.

Patch-set: 2