Update patch set 3
Patch Set 3:

Yes, exactly. Update all of them. We do not want to have to deal with some random dependency that we missed. You are updating socket.h, so update all the related headers.

In fact, now that I think about it, please make updating socket.h (and friends) a separate change. Then make this change sit on top of it that just includes the CAN stuff.

Patch-set: 3