Update patch set 1
Patch Set 1: Do not submit

patches to kernel-headers should always be exact copies of kernel headers. the check-in comment should include the SHA. typically the kernel header should come from the version of the kernel appropriate to the branch. (that would be the _device_ kernel, since the emulator kernel is typically a fossil.)

Patch-set: 1
Label: Code-Review=-2