Update patch set 1
Patch Set 1:

this is another instance of the same problem raghu reported. here we hacked the linux header file, but raghu didn't want to hack another header file, so we're fixing the script instead. once the script is fixed, we can remove this old hack, so now icmp.h really is an untouched kernel header file again.

Patch-set: 1