Update linux/netlink.h
From Linux kernel SHA 9baa5059485d6559081d27d9b1fe1714f3fccab1

This is an updated version of netlink.h from the Linux kernel.
In the kernel, this file is known as ./include/uapi/linux/netlink.h,
however, I renamed the file according to our messed up, old style
naming convention.

In addition, I changed the #ifdef at the top of the file from:

----- ./include/uapi/linux/netlink.h -----
...
----- end -----

to

----- ./original/linux/netlink.h -----
...
----- end -----

external/libnl-headers depends on the define being named __LINUX_NETLINK_H
and doesn't compile if we change the name.
This can be fixed in a followup CL.

Change-Id: I95d0876e1696ca265b3fa6c8b35c5c2d6eba4580