Update patch set 1
Patch Set 1:

we've already got too many manual hacks and i don't want to end up with the same unmaintainable mess we already have for 32-bit, so let's accept that fixing the script to work with the new headers is just part of the job of bringing in new headers. it seemed like we had a couple of plausible alternatives that mostly involved removing code, no?

Patch-set: 1