Update patch set 1
Patch Set 1:

it's never any single change. it's the accumulation of lots of tiny changes over time. we need to accept that our script's current model is broken and fix it, not pretend that we can keep randomly hacking supposedly pristine upstream files without paying for it later.

our job is supposed to be making things better, not worse.

Patch-set: 1