Update patch set 2
Patch Set 2:

can we get the i386, mips, and x86_64 headers too?

(are the x86_64 headers distinct from i386, or are they shared and just check __i386__/__LP64__?)

i'd particularly like to move x86_64 over to uapi because it's still currently easier for me to do the 64-bit work for x86_64 than aarch64, and no one but me is using x86_64 right now so it's not like it would be breaking anyone's build.

Patch-set: 2