Update patch set 1
Patch Set 1:

I thought about that, but this directory only really contains kernel header files, so it felt a bit wrong to stuff a script we wrote here. I'm not sure the bionic/libc/kernel/tools directory is perfect, but it feels like a good idea to keep all of our scripts in one place.

Patch-set: 1